YEEEHHHHHH ramp shows up

+++ Projectfiles/Scenes/Gameplay/Objects/Knight.h
@@ -41,7 +41,7 @@
 
 @property(nonatomic, assign) int jumpCounter;
 
-@property(nonatomic, assign) int fuel;
+@property(nonatomic, assign) float fuel;
 
 @property(nonatomic, assign) int coins;
